устранение замечаний ревью
