+++ app/src/main/java/ru/practicum/android/diploma/ui/favorite/FavoriteFragmentUpdate.kt
@@ -1,0 +1,14 @@
+package ru.practicum.android.diploma.ui.favorite
+
+import ru.practicum.android.diploma.domain.models.vacacy.Vacancy
+
+sealed class FavoriteFragmentUpdate {
+
+    data object EmptyVacancyList : FavoriteFragmentUpdate()
+
+    data object GetVacanciesError : FavoriteFragmentUpdate()
+
+    data class VacancyList(
+        val vacancies: List<Vacancy>
+    ) : FavoriteFragmentUpdate()
+}
